reformatted tf backend for mean

+++ ivy/functional/backends/tensorflow/statistical.py
@@ -79,7 +79,7 @@ def mean(
 ) -> Union[tf.Tensor, tf.Variable]:
     axis = tuple(axis) if isinstance(axis, list) else axis
     if dtype is not None:
-        x = x.astype(dtype)
+        x = tf.cast(x, dtype)
     else:
         dtype = x.dtype
     return tf.math.reduce_mean(x, axis=axis, keepdims=keepdims).astype(dtype)
